mpsl: cx: allow seamless adding of cx implementations
Removed mpsl_cx.c file and moved initialization of cx to
corresponding cx implementations. This allows adding of cx
implementations also as modules in customer-specific repositories
which is not referenced by the sdk-nrf repository.
