+++ subsys/mpsl/cx/bluetooth/mpsl_cx_1w_bluetooth.c
@@ -86,3 +86,12 @@ int mpsl_cx_bt_interface_1wire_config_set(void)
 
 	return err;
 }
+
+static int mpsl_cx_init(const struct device *dev)
+{
+	ARG_UNUSED(dev);
+
+	return mpsl_cx_bt_interface_1wire_config_set();
+}
+
+SYS_INIT(mpsl_cx_init, POST_KERNEL, CONFIG_KERNEL_INIT_PRIORITY_DEVICE);
